led-piano-kit/requirements.txt
mido
python-rtmidi
rpi-ws281x
adafruit-circuitpython-neopixel
adafruit-blinka
